Move callback code to it's own file
